+++ docs/release-notes.md
@@ -6,6 +6,7 @@
   * Fixed TrainerMod's `player_setlevel` command not also setting XP.
   * Renamed the default _TrainerMod_ mod to _Console Commands_ to clarify its purpose.
   * Added a log parser service at [log.smapi.io](https://log.smapi.io).
+  * Added better Steam instructions to the SMAPI installer.
 
 * For modders:
   * Added support for public code in reflection API, to simplify mod integrations.